+++ docs/constants/navigation.js
@@ -450,6 +450,7 @@ const preview = [
   makeSection('Preview', [
     makePage('preview/introduction.mdx'),
     makePage('preview/support.mdx'),
+    makePage('router/reference/not-found.mdx'),
     { expanded: true },
     makeSection('Expo Router', [makePage('preview/api-routes.mdx')]),
   ]),