+++ packages/@expo/cli/CHANGELOG.md
@@ -4,6 +4,16 @@
 
 ### 🛠 Breaking changes
 
+### 🎉 New features
+
+### 🐛 Bug fixes
+
+### 💡 Others
+
+## 0.14.0 — 2023-10-17
+
+### 🛠 Breaking changes
+
 - Enable `tsconfigPaths` by default flag ([#24490](https://github.com/expo/expo/pull/24490) by [@marklawlor](https://github.com/marklawlor))
 
 ### 🎉 New features